fixed bad initialization for md5

+++ include/ssl_dgst.h
@@ -45,7 +45,7 @@ int				ssl_dgst64_fillbuf(t_dgst64 *context, char const *src, \
 
 int				ssl_md5(t_arg *args, int flags);
 uint32_t		ssl_md5_decode(unsigned char *input);
-void			ssl_md5_encode(uint32_t n, unsigned char *s);
+void			ssl_md5_encode(uint64_t n, unsigned char *s);
 void			ssl_md5_update(t_dgst32 *context);
 void			ssl_md5_print(t_dgst32 *context, char *fname, \
 								int argflag, int *flags);
